+++ src/main/java/com/vomiter/rangedjs/mixin/BowAttributesOnReleaseMixin.java
@@ -5,16 +5,23 @@ import com.llamalad7.mixinextras.sugar.Local;
 import com.vomiter.rangedjs.item.bow.BowItemInterface;
 import com.vomiter.rangedjs.item.bow.BowUtils;
 import com.vomiter.rangedjs.projectile.ProjectileInterface;
+import net.minecraft.world.entity.LivingEntity;
+import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.entity.projectile.AbstractArrow;
+import net.minecraft.world.item.ArrowItem;
 import net.minecraft.world.item.BowItem;
 import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.enchantment.Enchantments;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Constant;
+import org.spongepowered.asm.mixin.injection.ModifyConstant;
 import org.spongepowered.asm.mixin.injection.ModifyVariable;
 
 @Mixin(value = BowItem.class)
 public abstract class BowAttributesOnReleaseMixin implements BowItemInterface {
+    @SuppressWarnings("unused")
     @Unique
     BowItem rjs$bowItem =(BowItem)(Object)this;
 
@@ -38,6 +45,14 @@ public abstract class BowAttributesOnReleaseMixin implements BowItemInterface {
         return f;
     }
 
+    @ModifyConstant(
+            method = "releaseUsing",
+            constant = @Constant(floatValue = 3.0F)
+    )
+    private float setInitialSpeed(float constant){
+        return this.getBowAttributes().getArrowSpeedScale();
+    }
+
 
 
     @ModifyVariable(
@@ -47,16 +62,30 @@ public abstract class BowAttributesOnReleaseMixin implements BowItemInterface {
     )
     private boolean setInfinity(boolean bl){
         return bl || this.getBowAttributes().isInfinity();
-    };
+    }
 
     @ModifyVariable(
             method = "releaseUsing",
             at = @At("STORE"),
             ordinal = 1
     )
-    private boolean setSpecialInfinity(boolean bl){
-        return bl || this.getBowAttributes().isSpecialInfinity();
-    };
+    private boolean setSpecialInfinity(
+            boolean bl,
+            @Local(argsOnly = true)
+            ItemStack bow,
+            @Local(ordinal = 1) ItemStack arrow,
+            @Local(argsOnly = true) LivingEntity player
+    ){
+        boolean returnValue = bl || this.getBowAttributes().isSpecialInfinity();
+        if(!(player instanceof Player)) return returnValue;
+        if(!returnValue && this.getBowAttributes().isInfinity()){
+            ItemStack simulatorBow = bow.copy();
+            simulatorBow.enchant(Enchantments.INFINITY_ARROWS, 1);
+            returnValue = (arrow.getItem() instanceof ArrowItem
+                    && ((ArrowItem)arrow.getItem()).isInfinite(arrow, simulatorBow, (Player)player));
+        }
+        return returnValue;
+    }
 
     @ModifyVariable(
             method = "releaseUsing",
@@ -65,7 +94,7 @@ public abstract class BowAttributesOnReleaseMixin implements BowItemInterface {
     )
     private int setBowPower(int i){
         return i + this.getBowAttributes().getPower();
-    };
+    }
 
     @ModifyVariable(
             method = "releaseUsing",
@@ -74,7 +103,7 @@ public abstract class BowAttributesOnReleaseMixin implements BowItemInterface {
     )
     private int setKnockback(int i){
         return i + this.getBowAttributes().getKnockBack();
-    };
+    }
 
     //TODO refactor this with mixin extra someday.
     @ModifyVariable(
@@ -89,5 +118,5 @@ public abstract class BowAttributesOnReleaseMixin implements BowItemInterface {
         if(this.getBowAttributes().isNoDamage()) arrow.setBaseDamage(0);
         ((ProjectileInterface)arrow).rangedjs$setHitBehavior(getHitBehavior());
         return arrow;
-    };
+    }
 }